refactor: 정렬 클래스 위치 이동
