+++ feature/diary/src/main/java/com/boostcamp/dreamteam/dreamdiary/feature/diary/home/tablist/DiaryHomeTabListUIState.kt
@@ -6,19 +6,3 @@ data class DiaryHomeTabListUIState(
     val diaries: List<DiaryUi> = listOf(),
     val loading: Boolean = false,
 )
-
-data class DiarySort(
-    val type: DiarySortType,
-    val order: DiarySortOrder,
-)
-
-enum class DiarySortType {
-    CREATED,
-    UPDATED,
-    SLEEP,
-}
-
-enum class DiarySortOrder {
-    ASC,
-    DESC,
-}
