+++ feature/diary/src/main/java/com/boostcamp/dreamteam/dreamdiary/feature/diary/home/tablist/DiaryListTab.kt
@@ -51,14 +51,17 @@ import androidx.compose.ui.unit.dp
 import androidx.paging.PagingData
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySort
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySortOrder
 import com.boostcamp.dreamteam.dreamdiary.designsystem.theme.DreamdiaryTheme
 import com.boostcamp.dreamteam.dreamdiary.feature.diary.R
 import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.component.DiaryCard
-import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.tablist.DiarySortOrder.ASC
-import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.tablist.DiarySortOrder.DESC
-import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.tablist.DiarySortType.CREATED
-import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.tablist.DiarySortType.SLEEP
-import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.tablist.DiarySortType.UPDATED
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySortOrder.ASC
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySortOrder.DESC
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySortType
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySortType.CREATED
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySortType.SLEEP
+import com.boostcamp.dreamteam.dreamdiary.core.domain.usecase.DiarySortType.UPDATED
 import com.boostcamp.dreamteam.dreamdiary.feature.diary.model.DiaryUi
 import com.boostcamp.dreamteam.dreamdiary.feature.diary.model.LabelUi
 import com.boostcamp.dreamteam.dreamdiary.feature.diary.model.PagingIndexKey
